give hint for subdomain creation
I totally forgot about `jx-staging` for the staging applications. So, in order to avoid some confusions among readers, I think it is useful to give an example of what to define for your domain.

Let me know if this is better suited as a normal paragraph or Info inset (which I think is what `I>` means)?